util: move validate function to 'validate' package
Having 'util' import 'validate' makes it very easy to hit import cycles,
e.g. if you try to use a 'util' helper function in validate or validate
tests.

Validation seems like it fits better in the validate package anyways.
