+++ config/v2_0/config.go
@@ -16,9 +16,9 @@ package v2_0
 
 import (
 	"github.com/coreos/ignition/config/errors"
-	"github.com/coreos/ignition/config/util"
 	"github.com/coreos/ignition/config/v1"
 	"github.com/coreos/ignition/config/v2_0/types"
+	"github.com/coreos/ignition/config/validate"
 	"github.com/coreos/ignition/config/validate/report"
 
 	json "github.com/ajeddeloh/go-json"
@@ -57,7 +57,7 @@ func Parse(rawConfig []byte) (types.Config, report.Report, error) {
 		return types.Config{}, report.Report{}, errors.ErrUnknownVersion
 	}
 
-	rpt := util.ValidateConfig(rawConfig, config)
+	rpt := validate.ValidateConfig(rawConfig, config)
 	if rpt.IsFatal() {
 		return types.Config{}, rpt, errors.ErrInvalid
 	}
