+++ config/v2_3_experimental/config.go
@@ -16,9 +16,9 @@ package v2_3_experimental
 
 import (
 	"github.com/coreos/ignition/config/errors"
-	"github.com/coreos/ignition/config/util"
 	"github.com/coreos/ignition/config/v2_2"
 	"github.com/coreos/ignition/config/v2_3_experimental/types"
+	"github.com/coreos/ignition/config/validate"
 	"github.com/coreos/ignition/config/validate/report"
 
 	json "github.com/ajeddeloh/go-json"
@@ -57,7 +57,7 @@ func Parse(rawConfig []byte) (types.Config, report.Report, error) {
 		return types.Config{}, report.Report{}, errors.ErrInvalid
 	}
 
-	rpt := util.ValidateConfig(rawConfig, config)
+	rpt := validate.ValidateConfig(rawConfig, config)
 	if rpt.IsFatal() {
 		return types.Config{}, rpt, errors.ErrInvalid
 	}
